Fix progress status route's return content
At some point D2 stopped accepting Dancer2::Core::Response objects as
the value returned by a route.
You now need to modify the existing response object stored in the context.

The reason we need to do this is because we need to force the output to
json, regardless of whatever the App's settings are.